fix(core): clamp ol start instead of overflowing the counter
start was parsed as usize and the per-item increment was unchecked, so
start="18446744073709551615" with a single <li> overflowed and panicked -
failing the whole conversion, not just the list. The counter is now i64, so a
negative start counts down as browsers do, out-of-range magnitudes clamp with
a warn, and the increment pins at i64::MAX rather than wrapping. The
saturation warn fires once per list, not once per item, so a long list cannot
flood the logs.

+++ crates/html-to-markdown/src/converter/context.rs
@@ -37,8 +37,11 @@ pub type ImageMetadataPayload = (BTreeMap<String, String>, Option<u32>, Option<u
 pub struct Context {
     /// Are we inside a code-like element (pre, code, kbd, samp)?
     pub(crate) in_code: bool,
-    /// Current list item counter for ordered lists
-    pub(crate) list_counter: usize,
+    /// Current list item counter for ordered lists.
+    ///
+    /// Signed so a negative `start` attribute (permitted by the HTML spec and honored by
+    /// browsers) can be represented without a separate sign flag.
+    pub(crate) list_counter: i64,
     /// Are we in an ordered list (vs unordered)?
     pub(crate) in_ordered_list: bool,
     /// Blockquote nesting depth
